```julia
function foo()
    # This tests the MaterialParameters structure
    g = CharUnits_GEO = GEO_units(; viscosity=1e19, length=1000km)

    # Define a struct for a first phase
    x = SetMaterialParams(;
        Name="test1",
        Phase=22,
        CreepLaws=(PowerlawViscous(), LinearViscous(; η=1e21Pa * s)),
        Gravity=ConstantGravity(; g=11.0m / s^2),
        Density=ConstantDensity(),
    )

    x_nd = nondimensionalize(x, CharUnits_GEO)

    # Define a struct for a first phase
    y = SetMaterialParams(;
        Name="test1",
        Phase=22,
        CreepLaws=(PowerlawViscous(), LinearViscous(; η=1e21Pa * s)),
        Gravity=ConstantGravity(; g=11.0m / s^2),
        Density=ConstantDensity(),
        CharDim = CharUnits_GEO
    )

    a, b, c = 0, 0, 0
    ta, tb, tc = 0, 0, 0
    for _ in 1:1_000_000
        ta += @elapsed begin a += @allocated begin compute_density(x_nd, (;)) end end
        tb += @elapsed begin b += @allocated begin compute_density(x, (;)) end end
        tc += @elapsed begin c += @allocated begin compute_density(y, (;)) end end
    end

    println("a => $a allocated, $ta seconds")
    println("b => $b allocated, $tb seconds")
    println("c => $c allocated, $tc seconds")
end
```

This PR
```julia-repl
In [16]: foo()
a => 0 allocated, 0.0568452000006039 seconds
b => 0 allocated, 0.05718950000061417 seconds
c => 0 allocated, 0.05717680000061389 seconds
```

#main
```julia-repl
In [1]: foo()
a => 16000000 allocated, 0.08244640000127741 seconds
b => 0 allocated, 0.0606366000007116 seconds
c => 16000000 allocated, 0.08222860000127924 seconds
```

So theres a c. 40% performance improvement and no more allocations :)